requirements.txt
2captcha-python==1.1.3
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
revChatGPT==0.3.1
selenium==4.8.0
slack-sdk==3.19.5
sniffio==1.3.0
sortedcontainers==2.4.0
tls-client==0.1.8
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
